+++ src/VersionModule.f90
@@ -1,4 +1,4 @@
 module VersionModule
     implicit none
-    character(len=17), parameter :: MODEL_VERSION = "0.0.4-16-gd0db59a"
+    character(len=17), parameter :: MODEL_VERSION = "0.0.4-19-g674d437"
 end module